[#FIX #1387] Add a timeout to validation messages.
Currently when a validation error occurs it stays indefinitely and
takes priority over any other window actions (this means the keyboard
is dismissed and any of the pickers for a particular command
disappear)

This also means that after we've popped a validation error, the
keyboard won't come back up unless we explicitly tap the input field
rather than the command in the drawer.

In the error-events-creator, we can set a timeout on validation
messages so they are dismissed after x amount of time (I've
arbitrarily chosen 2 seconds). After the error message is dimissed,
the keyboard responds as it should.